Include the correct headers in type/deprecated/constant_sequence.h
Because the code is dependent on `std::char_traits` and `std::allocator` being defined, but, if this header is included before `<string>` and `<memory>` are included, they won't be, causing the build to fail.
This is the case when building Thrift on OSX, and was blocking HHVM from being able to update to the newest version of Thrift.