Add missing namespace anchor in data-member getters
The following example code can trigger a compilation failure:

```lang=c++
struct xyz {
  // not too uncommon to use the name `std` in statistics-related code
  FATAL_DATA_MEMBER_GETTER(std, std);
  FATAL_DATA_MEMBER_GETTER(abc, abc);
}
```

The problem is that there is a single occurrence in the macro
implementing `FATAL_DATA_MEMBER_GETTER` using `std` without `::` to its
left. In the example, in the first invocation, the local name `std` gets
defined. In the second invocation, the solitary unqualified use of the
name `std` becomes ambiguous with the namespace `std`. The intention is
to use various traits types in the `std` namespace.