Fix UBSAN failures in fatal/type/test/enum_test.cpp
Summary:
In that test, we have casts from integral values to unscoped enum types, where the values are outside of the range of the casted-to unscoped enum types.

The fixes are either to use scoped enum types only or to used unscoped enum types with fixed underlying types only. However, it is straightforward to use scoped enum types only because that involves a simple substitution of one macro for another, so we choose this fix.

For unscoped enum types without fixed underlying types, the range is specified as:
> all values possible for the smallest bit field large enough to hold all enumerators of the target enumeration.
> http://en.cppreference.com/w/cpp/language/enum

The following tests in `fatal/type/test/enum_test.cpp` were failing when run with Undefined Behavior Sanitizer:
```
enums - is_valid
enums - to_string
enums - try_parse
enums - is_valid_enum
enums - enum_to_string
```

Where the following class of error appears:
```
runtime error: load of value <X>, which is not a valid value for type '<Y>'
```

The actual specification is more complicated than appears above, and can be found at:
> C++14 draft standard, 7.2 [dcl.enum] #8, http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2014/n4296.pdf

Reviewed By: andrewjcg

Differential Revision: D4857422

fbshipit-source-id: 1d3c26dc0d71c937f08a6048a32c59a44293c852