Simplify implementations of get, find, and contains
Summary:
[Fatal] Simplify implementations of `get`, `find`, and `contains`.

The existing implementations have explicit variants for each variation of filter arguments. But this can also be done simply by providing the right defaults to the most complex variants and removing the others.

Assumes the identity template alias is nearly free in terms of compilation time, lest this be a regression.

Reviewed By: ericniebler

Differential Revision: D5018345

fbshipit-source-id: 90c44d4928ec8c1e2b23236335fe0397404ee4f3