[Fatal] Alias fatal::sequence to std::integer_sequence
Alias `fatal::sequence` to `std::integer_sequence` when the platform has it. And alias `fatal::make_sequence` to `std::make_integer_sequence` when the platform has a reasonably efficient implementation of it.

Useful because Clang's `std::make_integer_sequence` is implemented as a compiler intrinsic: https://reviews.llvm.org/D13786. And supposedly MSVC's is also an intrinsic (unverified).