let test printers bind to their outputs
Summary: Rather than passing outputs all the way through, templated. The out param is always some kind of `std::ostream` and there is no overpowering need to hide that fact. The out param is the same for every call to the same instance of a printer, and there is no need to hide that fact either.

Reviewed By: Gownta

Differential Revision: D37355868

fbshipit-source-id: 1ffdf265cbce7d4a4a39dae147992202e79530d5