pre-shuffle in sort
Summary:
It is common to apply the sort algorithm to lists which are already sorted or mostly-sorted. The sort algorithm uses quicksort, sorts into inverse order, and chooses the first element as the pivot. For already sorted or mostly-sorted lists, this yields pathological quadratic behavior v.s. the log-linear expected behavior and linear recursion depth v.s. the log expected recursion depths. This is observed in excessive compile times, excessive memory consumption at compile time, and failures due to exceeding the template recursion depth or lack of available memory.

Taking all lists uniformly at random, no way of choosing a pivot does better than any other. But we do not take all lists uniformly at random - it is common for lists to be sorted or almost-sorted. So we select an approach taking into account that we do not take all lists uniformly at random.

Because this happens at compile time, there is no risk worth worrying about of malicious input that can specifically target the selected approach.

There are various possible solutions. This solution pre-shuffles the list. As a consequence, the sort is not stable.

Reviewed By: Alfus

Differential Revision: D37254883

fbshipit-source-id: fbbf9556ce0cbb6b16923d486486406a2555cfbb