revise partition algorithm in sort onto cat
Summary:
There is a partition algorithm used by `sort` which attempts to optimize up to 3-at-once. Optimizing the partition algorithm directly is much more cumbersome because it must be exponential in specializations - 2^N rather than N - for optimizing N-at-once.

Revise the algorithm not to optimize N-at-once but instead just to forward on to `cat` since `cat` can optimize N-at-once with N specializations rather than 2^N specializations and does so, with much larger N than the partition algorithm can do.

This causes a small regression in compile time and peak-rss, but in this case the win in code complexity outweighs the price.

Reviewed By: Gownta

Differential Revision: D39745818

fbshipit-source-id: 61feeb68dc5b6503be729025129162eecf9763c3