Add json output capabilities to fatal/ gtest framework
Summary:
the *fatal* project owners created a mini cpp framework for their folder which doesn't have some implementation of the gtest framework e.g writing the listing output in json to a file.

This change implements writing json output to a file for the fatal gtest framework during test execution.

NOTE: Ideally, this needs to be it's own framework and not be considered gtest from within tpx as this isn't totally compatible.

Reviewed By: patskovn

Differential Revision: D77318081

fbshipit-source-id: 7d4da510af03d101e4bd3223d40d7fd0532eb150